chore: reorganize project
