+++ .gitignore
@@ -1,3 +1,5 @@
 *.log
 *.pyc
 .git
+
+/dist
