+++ pyproject.toml
@@ -1,0 +1,30 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "dlmyo"
+version = "0.0.1"
+authors = [
+  { name="Max Leefer" },
+  { name="Siarhei Yorsh" },
+  { name="Iori Mizutani", email="iori.mizutani@gmail.com" },
+]
+Maintainers = [
+  { name="Iori Mizutani", email="iori.mizutani@gmail.com" },
+]
+description = "A middleware to manage Myo EMG streams"
+readme = "README.md"
+requires-python = ">=3.11"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: POSIX :: Linux",
+    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
+    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/iomz/dlmyo"
+"Bug Tracker" = "https://github.com/iomz/dlmyo/issues"
+
+[tool]
