+++ src/dlmyo/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+Dongleless Myo EMG Streamer
+
+Authors:
+   - Max Leefer (@mamo91)
+   - Siarhei Yorsh (@MyrikLD)
+   - Iori Mizutani (@iomz)
+
+~~~~~~~~~~~~~~~~~~~~
+   >>> import dlmyo
+"""
+from .dongleless import *
